Add validator check to ensure a Wikidata item exists and is no redirect
This validator check needs a network connection, if the API requests fail, this is noted in the 'Other' category of the validator.
It provides an auto-fix if the wikidata=* tag points to a redirect. If the Wikidata item does not exist, the users have to fix it on their own.
The class AllValidatorTests will serve as a directory where all possible TestErrors are listed.