Make use of caching for certain API requests
At the moment it's only the `action=sitematrix` API call that will be cached (for 30 days).
The queries for which caching is enabled will be stored in the cache for a certain amount of time (configured per query). After that time passed, the query gets sent to the API again. If that request fails, the cached result will be used even longer. But if it's successful, the cache is updated and the timer restarts.