No longer use `CapturingInputStream` and `InvalidApiQueryException`
The former is replaced by first reading the String from the stream, only then start decoding the JSON.
The latter is now just a simple IOException, that exception was no longer up to date, because for POST requests the query is no longer in the URL, so that Exception would not contain the important information (i.e. the query).