Provide an auto-fix for wikidata=* tags with a brand as value
If there is a item about a brand in the wikidata=* tag, there is now an auto-fix that adds the value to the brand:wikidata=* tag. If the value is already present in the latter tag, the wikidata=* tag is simply dropped.
This fixes https://gitlab.com/JOSM/wikipedia/issues/2 , where this was proposed.