Let Ivy exclude test dependencies
We hit the issue described in https://github.com/Wikidata/Wikidata-Toolkit/issues/495 , that the transitive dependency on `org.wikidata.wdtk:wdtk-wikibaseapi:0.11.1` did not exist. But since it is only a dependency with Maven dependency scope `test` (in Ivy terminology "conf=test"), we should be able to work around it by only pulling in dependencies of the default scope.